[DF-7307] Using a built-in integer type instead DAVA::int32
Former-commit-id: d4577554e0b98f0bae9f4be97fc3533ea34fb6d5

+++ Sources/Internal/Concurrency/Spinlock.h
@@ -51,7 +51,7 @@ public:
     void Unlock() DAVA_NOEXCEPT;
 
 private:
-    Atomic<int32> flag;
+    Atomic<size_t> flag;
 };
 
 //-----------------------------------------------------------------------------
